Move out AST rewriting.
